Remove `Finder::Context` module and corresponding directory.
